# fix: refactor charts for validation SQL to use a cte instead of sub q…