--- a/packages/backend/src/models/SavedChartModel.ts
+++ b/packages/backend/src/models/SavedChartModel.ts
@@ -980,97 +980,109 @@ export class SavedChartModel {
             sorts: string[];
         }>
     > {
-        return this.database(SavedChartsTableName)
-            .select({
-                uuid: 'saved_queries.saved_query_uuid',
-                name: 'saved_queries.name',
-                tableName: 'saved_queries_versions.explore_name',
-                filters: 'saved_queries_versions.filters',
-                dimensions: this.database.raw(
-                    "COALESCE(ARRAY_AGG(DISTINCT svf.name) FILTER (WHERE svf.field_type = 'dimension'), '{}')",
-                ),
-                metrics: this.database.raw(
-                    "COALESCE(ARRAY_AGG(DISTINCT svf.name) FILTER (WHERE svf.field_type = 'metric'), '{}')",
-                ),
-                tableCalculations: this.database.raw(
-                    "COALESCE(ARRAY_AGG(DISTINCT sqvtc.name) FILTER (WHERE sqvtc.name IS NOT NULL), '{}')",
-                ),
-                customMetrics: this.database.raw(
-                    "COALESCE(ARRAY_AGG(DISTINCT (sqvam.table || '_' || sqvam.name)) FILTER (WHERE sqvam.name IS NOT NULL), '{}')",
-                ),
-                customMetricsBaseDimensions: this.database.raw(
-                    "COALESCE(ARRAY_AGG(DISTINCT (sqvam.table || '_' || sqvam.base_dimension_name)) FILTER (WHERE sqvam.base_dimension_name IS NOT NULL), '{}')",
-                ),
-                customBinDimensions: this.database.raw(
-                    "COALESCE(ARRAY_AGG(DISTINCT sqvcd.id) FILTER (WHERE sqvcd.id IS NOT NULL), '{}')",
-                ),
-                customSqlDimensions: this.database.raw(
-                    "COALESCE(ARRAY_AGG(DISTINCT sqvcsd.id) FILTER (WHERE sqvcsd.id IS NOT NULL), '{}')",
-                ),
-                sorts: this.database.raw(
-                    "COALESCE(ARRAY_AGG(DISTINCT sqvs.field_name) FILTER (WHERE sqvs.field_name IS NOT NULL), '{}')",
-                ),
-            })
-            .leftJoin(
-                SpaceTableName,
-                'saved_queries.space_id',
-                'spaces.space_id',
-            )
-            .leftJoin(
-                ProjectTableName,
-                'spaces.project_id',
-                'projects.project_id',
-            )
-            .leftJoin(
-                SavedChartVersionsTableName,
-                'saved_queries.saved_query_id',
-                'saved_queries_versions.saved_query_id',
-            )
-            .leftJoin(
-                'saved_queries_version_fields as svf',
-                'saved_queries_versions.saved_queries_version_id',
-                'svf.saved_queries_version_id',
-            )
-            .leftJoin(
-                'saved_queries_version_table_calculations as sqvtc',
-                'saved_queries_versions.saved_queries_version_id',
-                'sqvtc.saved_queries_version_id',
-            )
-            .leftJoin(
-                'saved_queries_version_additional_metrics as sqvam',
-                'saved_queries_versions.saved_queries_version_id',
-                'sqvam.saved_queries_version_id',
-            )
-            .leftJoin(
-                'saved_queries_version_custom_dimensions as sqvcd',
-                'saved_queries_versions.saved_queries_version_id',
-                'sqvcd.saved_queries_version_id',
-            )
-            .leftJoin(
-                'saved_queries_version_custom_sql_dimensions as sqvcsd',
-                'saved_queries_versions.saved_queries_version_id',
-                'sqvcsd.saved_queries_version_id',
-            )
-            .leftJoin(
-                'saved_queries_version_sorts as sqvs',
-                'saved_queries_versions.saved_queries_version_id',
-                'sqvs.saved_queries_version_id',
-            )
-            .where(
-                'saved_queries_versions.saved_queries_version_id',
-                this.database('saved_queries_versions')
-                    .select('saved_queries_version_id')
-                    .where(
-                        'saved_queries_versions.saved_query_id',
-                        this.database
-                            .ref('saved_query_id')
-                            .withSchema(SavedChartsTableName),
-                    )
-                    .orderBy('saved_queries_versions.created_at', 'desc')
-                    .limit(1),
-            )
-            .andWhere('projects.project_uuid', projectUuid)
-            .groupBy(1, 2, 3, 4);
+        const cteName = 'chart_last_version_cte';
+        return (
+            this.database
+                // cte to get the last version of each chart in the project
+                .with(cteName, (qb) => {
+                    void qb
+                        .select({
+                            saved_query_uuid: 'saved_queries.saved_query_uuid',
+                            name: 'saved_queries.name',
+                            saved_queries_version_id: this.database.raw(
+                                'MAX(saved_queries_versions.saved_queries_version_id)',
+                            ),
+                        })
+                        .from(SavedChartsTableName)
+                        .leftJoin(
+                            SpaceTableName,
+                            'saved_queries.space_id',
+                            'spaces.space_id',
+                        )
+                        .leftJoin(
+                            ProjectTableName,
+                            'spaces.project_id',
+                            'projects.project_id',
+                        )
+                        .leftJoin(
+                            SavedChartVersionsTableName,
+                            'saved_queries.saved_query_id',
+                            'saved_queries_versions.saved_query_id',
+                        )
+                        .where('projects.project_uuid', projectUuid)
+                        .groupBy(
+                            'saved_queries.saved_query_uuid',
+                            'saved_queries.name',
+                        );
+                })
+                .select({
+                    uuid: `${cteName}.saved_query_uuid`,
+                    name: `${cteName}.name`,
+                    tableName: 'saved_queries_versions.explore_name',
+                    filters: 'saved_queries_versions.filters',
+                    dimensions: this.database.raw(
+                        "COALESCE(ARRAY_AGG(DISTINCT svf.name) FILTER (WHERE svf.field_type = 'dimension'), '{}')",
+                    ),
+                    metrics: this.database.raw(
+                        "COALESCE(ARRAY_AGG(DISTINCT svf.name) FILTER (WHERE svf.field_type = 'metric'), '{}')",
+                    ),
+                    tableCalculations: this.database.raw(
+                        "COALESCE(ARRAY_AGG(DISTINCT sqvtc.name) FILTER (WHERE sqvtc.name IS NOT NULL), '{}')",
+                    ),
+                    customMetrics: this.database.raw(
+                        "COALESCE(ARRAY_AGG(DISTINCT (sqvam.table || '_' || sqvam.name)) FILTER (WHERE sqvam.name IS NOT NULL), '{}')",
+                    ),
+                    customMetricsBaseDimensions: this.database.raw(
+                        "COALESCE(ARRAY_AGG(DISTINCT (sqvam.table || '_' || sqvam.base_dimension_name)) FILTER (WHERE sqvam.base_dimension_name IS NOT NULL), '{}')",
+                    ),
+                    customBinDimensions: this.database.raw(
+                        "COALESCE(ARRAY_AGG(DISTINCT sqvcd.id) FILTER (WHERE sqvcd.id IS NOT NULL), '{}')",
+                    ),
+                    customSqlDimensions: this.database.raw(
+                        "COALESCE(ARRAY_AGG(DISTINCT sqvcsd.id) FILTER (WHERE sqvcsd.id IS NOT NULL), '{}')",
+                    ),
+                    sorts: this.database.raw(
+                        "COALESCE(ARRAY_AGG(DISTINCT sqvs.field_name) FILTER (WHERE sqvs.field_name IS NOT NULL), '{}')",
+                    ),
+                })
+                .from(cteName)
+                .leftJoin(
+                    SavedChartVersionsTableName,
+                    `${cteName}.saved_queries_version_id`,
+                    'saved_queries_versions.saved_queries_version_id',
+                )
+                .leftJoin(
+                    'saved_queries_version_fields as svf',
+                    'saved_queries_versions.saved_queries_version_id',
+                    'svf.saved_queries_version_id',
+                )
+                .leftJoin(
+                    'saved_queries_version_table_calculations as sqvtc',
+                    'saved_queries_versions.saved_queries_version_id',
+                    'sqvtc.saved_queries_version_id',
+                )
+                .leftJoin(
+                    'saved_queries_version_additional_metrics as sqvam',
+                    'saved_queries_versions.saved_queries_version_id',
+                    'sqvam.saved_queries_version_id',
+                )
+                .leftJoin(
+                    'saved_queries_version_custom_dimensions as sqvcd',
+                    'saved_queries_versions.saved_queries_version_id',
+                    'sqvcd.saved_queries_version_id',
+                )
+                .leftJoin(
+                    'saved_queries_version_custom_sql_dimensions as sqvcsd',
+                    'saved_queries_versions.saved_queries_version_id',
+                    'sqvcsd.saved_queries_version_id',
+                )
+                .leftJoin(
+                    'saved_queries_version_sorts as sqvs',
+                    'saved_queries_versions.saved_queries_version_id',
+                    'sqvs.saved_queries_version_id',
+                )
+                .groupBy(1, 2, 3, 4)
+        );
     }
 
     async find(filters: {