…uery

### Description:

Related to comment from previous PR. https://github.com/lightdash/lightdash/pull/10226#discussion_r1618760457
> this subselect can be the new bottleneck, but needed to the the latest version. let's keep eye.

SQL optimization: moved from using subquery to use a cte

Before:

> 417 rows retrieved starting from 1 in 863 ms (execution: 565 ms, fetching: 298 ms)

```
SELECT sq.saved_query_uuid,
       sq.name,
       sv.explore_name,
       sv.filters,
       coalesce(ARRAY_AGG(DISTINCT svf.name) FILTER (WHERE svf.field_type = 'dimension'), '{}') as dimensions,
       coalesce(ARRAY_AGG(DISTINCT svf.name) FILTER (WHERE svf.field_type = 'metric'), '{}')    as metrics,
       coalesce(ARRAY_AGG(DISTINCT sqvtc.name) FILTER (WHERE sqvtc.name IS NOT NULL), '{}')     as table_calculations,
       coalesce(ARRAY_AGG(DISTINCT sqvam.name) FILTER (WHERE sqvam.name IS NOT NULL), '{}')     as custom_metrics,
       coalesce(ARRAY_AGG(DISTINCT sqvcd.id) FILTER (WHERE sqvcd.id IS NOT NULL),
                '{}')                                                                           as custom_bin_dimensions,
       coalesce(ARRAY_AGG(DISTINCT sqvcsd.id) FILTER (WHERE sqvcsd.id IS NOT NULL),
                '{}')                                                                           as custom_sql_dimensions,
       coalesce(ARRAY_AGG(DISTINCT sqvs.field_name) FILTER (WHERE sqvs.field_name IS NOT NULL),
                '{}')                                                                           as sorts
FROM saved_queries as sq
         LEFT JOIN spaces as s
                   ON sq.space_id = s.space_id
         LEFT JOIN projects as p
                   ON s.project_id = p.project_id
         LEFT JOIN saved_queries_versions as sv
                   ON sq.saved_query_id = sv.saved_query_id
         LEFT JOIN saved_queries_version_fields as svf
                   ON sv.saved_queries_version_id = svf.saved_queries_version_id
         LEFT JOIN saved_queries_version_table_calculations as sqvtc
                   ON sv.saved_queries_version_id = sqvtc.saved_queries_version_id
         LEFT JOIN saved_queries_version_additional_metrics as sqvam
                   ON sv.saved_queries_version_id = sqvam.saved_queries_version_id
         LEFT JOIN saved_queries_version_custom_dimensions as sqvcd
                   ON sv.saved_queries_version_id = sqvcd.saved_queries_version_id
         LEFT JOIN saved_queries_version_custom_sql_dimensions as sqvcsd
                   ON sv.saved_queries_version_id = sqvcsd.saved_queries_version_id
         LEFT JOIN saved_queries_version_sorts as sqvs
                   ON sv.saved_queries_version_id = sqvs.saved_queries_version_id
WHERE sv.saved_queries_version_id = (SELECT saved_queries_version_id
                                     FROM saved_queries_versions
                                     WHERE saved_queries_versions.saved_query_id = sq.saved_query_id
                                     ORDER BY saved_queries_versions.created_at desc
                                     LIMIT 1)
  AND p.project_uuid = ?
GROUP BY 1, 2, 3, 4
```

After:

> 417 rows retrieved starting from 1 in 715 ms (execution: 543 ms, fetching: 172 ms)

```
with cte as
         (SELECT sq.saved_query_uuid,
                 sq.name,
                 MAX(sv.saved_queries_version_id) as saved_queries_version_id
          FROM saved_queries as sq
                   LEFT JOIN spaces as s
                             ON sq.space_id = s.space_id
                   LEFT JOIN projects as p
                             ON s.project_id = p.project_id
                   LEFT JOIN saved_queries_versions as sv
                             ON sq.saved_query_id = sv.saved_query_id
          WHERE p.project_uuid = ?
          GROUP BY 1, 2)

SELECT cte.saved_query_uuid,
       cte.name,
       sv.explore_name,
       sv.filters,
       coalesce(ARRAY_AGG(DISTINCT svf.name) FILTER (WHERE svf.field_type = 'dimension'), '{}') as dimensions,
       coalesce(ARRAY_AGG(DISTINCT svf.name) FILTER (WHERE svf.field_type = 'metric'), '{}')    as metrics,
       coalesce(ARRAY_AGG(DISTINCT sqvtc.name) FILTER (WHERE sqvtc.name IS NOT NULL), '{}')     as table_calculations,
       coalesce(ARRAY_AGG(DISTINCT sqvam.name) FILTER (WHERE sqvam.name IS NOT NULL), '{}')     as custom_metrics,
       coalesce(ARRAY_AGG(DISTINCT sqvcd.id) FILTER (WHERE sqvcd.id IS NOT NULL),
                '{}')                                                                           as custom_bin_dimensions,
       coalesce(ARRAY_AGG(DISTINCT sqvcsd.id) FILTER (WHERE sqvcsd.id IS NOT NULL),
                '{}')                                                                           as custom_sql_dimensions,
       coalesce(ARRAY_AGG(DISTINCT sqvs.field_name) FILTER (WHERE sqvs.field_name IS NOT NULL),
                '{}')                                                                           as sorts
FROM cte
         LEFT JOIN saved_queries_versions as sv
                   ON cte.saved_queries_version_id = sv.saved_queries_version_id
         LEFT JOIN saved_queries_version_fields as svf
                   ON cte.saved_queries_version_id = svf.saved_queries_version_id
         LEFT JOIN saved_queries_version_table_calculations as sqvtc
                   ON cte.saved_queries_version_id = sqvtc.saved_queries_version_id
         LEFT JOIN saved_queries_version_additional_metrics as sqvam
                   ON cte.saved_queries_version_id = sqvam.saved_queries_version_id
         LEFT JOIN saved_queries_version_custom_dimensions as sqvcd
                   ON cte.saved_queries_version_id = sqvcd.saved_queries_version_id
         LEFT JOIN saved_queries_version_custom_sql_dimensions as sqvcsd
                   ON cte.saved_queries_version_id = sqvcsd.saved_queries_version_id
         LEFT JOIN saved_queries_version_sorts as sqvs
                   ON cte.saved_queries_version_id = sqvs.saved_queries_version_id
GROUP BY 1, 2, 3, 4
```

### Reviewer actions

- [ ] I have manually tested the changes in the preview environment
- [ ] I have reviewed the code
- [ ] I understand that "request changes" will block this PR from merging